Abstract classes
Partially implemented an abstract class to represent the scheduler, along with LRU and Clock policy classes that extend scheduler. The only difference between the two is the algorithm they use to replace pages. May need to implement a group of arrays or arraylists to represent main memory, the sizes of which can be determined by dividing up the input value numFrames, between the processes. Could then "load" a page into an array index (frame). Need to think about a bit value to use with each "Frame" in the clock policy.