LRU mostly done
LRU seems to be working ok. Nest up is to implement an event log output. As it is, it might pay to have the algorithms in separate classes after all. No time to implement inheritance at the moment. So will copy then refactor the scheduler class. i.e. schedulerLRU and schedulerClock