Implemented Clock scheduling policy
Implemented the clock scheduling policy. It's pretty much a copy of the LRU class but has a different swapping algorithm and uses the marker arraylist to store bit values instead of time stamps. If i had more time i would have liked to convert them into inherited classes. 
Implemented a circular arraylist iterator to make the marker list behave as required. 
The algorithms are finishing too soon. This is happening in both classes, so it it something common to both of them.